Move all activities into their own package
This helps organizes the files a bit better.
ListEntriesRecyclerView.java has been renamed to be more accurate,
along with renaming its class to ListEntriesRecyclerViewAdapter and making
all package-private methods to be public for Codacy.
