+++ frontend/app/src/main/java/com/quickpick/activities/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.quickpick;
+package com.quickpick.activities;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -15,6 +15,7 @@ import com.facebook.login.LoginResult;
 import com.facebook.login.widget.LoginButton;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.messaging.FirebaseMessaging;
+import com.quickpick.R;
 import com.quickpick.apis.LoginApi;
 import com.quickpick.apis.RetrofitUtils;
 import com.quickpick.payloads.LoginRequest;
