+++ frontend/app/src/main/res/layout/activity_session.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SessionActivity">
+    tools:context=".activities.SessionActivity">
 
     <Button
         android:id="@+id/start_swiping_button"
