+++ frontend/app/src/main/res/layout/activity_view_old_sesssions.xml
@@ -1,9 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ViewOldSessionsActivity">
+    tools:context=".activities.ViewOldSessionsActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
