+++ frontend/app/src/main/res/layout/activity_view_or_update_lists.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ViewOrUpdateListsActivity">
+    tools:context=".activities.ViewOrUpdateListsActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/list_recycler_view"
